requirements.txt
paddlepaddle==3.2.0
paddleocr==2.7.3
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow>=10.0.0
streamlit>=1.28.0
opencv-python<=4.6.0.66
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
python-Levenshtein>=0.21.0
datasets>=2.14.0
pandas>=1.5.0